refactor(infra): inline the warp check wrapper
