+++ typescript/infra/scripts/check/check-warp-deploy.ts
@@ -1,26 +1,50 @@
 import chalk from 'chalk';
 import { Gauge, Registry } from 'prom-client';
+import { stringify as yamlStringify } from 'yaml';
 
 import { submitMetrics } from '@hyperlane-xyz/metrics';
 import { getRegistry } from '@hyperlane-xyz/registry/fs';
-import { ChainName } from '@hyperlane-xyz/sdk';
+import {
+  type ChainName,
+  type WarpCoreConfig,
+  type WarpRouteCheckResult,
+  type WarpRouteDeployConfigMailboxRequired,
+  checkWarpRouteDeployConfig,
+} from '@hyperlane-xyz/sdk';
+import { assert, objFilter } from '@hyperlane-xyz/utils';
 
 import { WarpRouteIds } from '../../config/environments/mainnet3/warp/warpIds.js';
 import { DEFAULT_REGISTRY_URI } from '../../config/registry.js';
 import {
   getWarpConfigMap,
+  getWarpDeployConfigFromMergedRegistry,
   getWarpConfigMapFromMergedRegistry,
 } from '../../config/warp.js';
+import { type EnvironmentConfig } from '../../src/config/environment.js';
 import { getEnvironmentConfig } from '../core-utils.js';
 
 import {
   getCheckWarpDeployArgs,
   getCheckerViolationsGaugeObj,
 } from './check-utils.js';
-import {
-  logWarpRouteCheckResult,
-  runWarpRouteCheckFromRegistry,
-} from './check-warp-route.js';
+
+const ROUTES_TO_SKIP: string[] = [
+  WarpRouteIds.ArbitrumBaseBlastBscEthereumGnosisLiskMantleModeOptimismPolygonScrollZeroNetworkZoraMainnet,
+  'EDGEN/bsc-edgenchain-ethereum',
+  'INJ/inevm-injective',
+  'USDC/ethereum-inevm',
+  'USDT/ethereum-inevm',
+  'WBTC/ethereum-form',
+  'WSTETH/ethereum-form',
+  'USDT/ethereum-form',
+  'USDC/ethereum-form',
+  'TRUMP/arbitrum-avalanche-base-flowmainnet-form-optimism-solanamainnet-worldchain',
+  'AIXBT/base-form',
+  'FORM/ethereum-form',
+  'GAME/base-form',
+  // Skip until Paradex executes hyperevm upgrade on their side
+  WarpRouteIds.ParadexUSDC,
+];
 
 async function main() {
   const { environment, chains, pushMetrics } =
@@ -34,24 +58,6 @@ async function main() {
 
   const failedWarpRoutesChecks: string[] = [];
 
-  const routesToSkip: string[] = [
-    WarpRouteIds.ArbitrumBaseBlastBscEthereumGnosisLiskMantleModeOptimismPolygonScrollZeroNetworkZoraMainnet,
-    'EDGEN/bsc-edgenchain-ethereum',
-    'INJ/inevm-injective',
-    'USDC/ethereum-inevm',
-    'USDT/ethereum-inevm',
-    'WBTC/ethereum-form',
-    'WSTETH/ethereum-form',
-    'USDT/ethereum-form',
-    'USDC/ethereum-form',
-    'TRUMP/arbitrum-avalanche-base-flowmainnet-form-optimism-solanamainnet-worldchain',
-    'AIXBT/base-form',
-    'FORM/ethereum-form',
-    'GAME/base-form',
-    // Skip until Paradex executes hyperevm upgrade on their side
-    WarpRouteIds.ParadexUSDC,
-  ];
-
   const registries = [DEFAULT_REGISTRY_URI];
   const registry = getRegistry({
     registryUris: registries,
@@ -62,70 +68,23 @@ async function main() {
     await getWarpConfigMapFromMergedRegistry(registries);
 
   console.log(chalk.yellow('Skipping the following warp routes:'));
-  routesToSkip.forEach((route) => console.log(chalk.yellow(`- ${route}`)));
-
-  const isTestnetRoute = async (warpRouteConfig: any) => {
-    for (const chain of Object.keys(warpRouteConfig)) {
-      const chainMetadata = await registry.getChainMetadata(chain);
-      if (chainMetadata?.isTestnet) return true;
-    }
-    return false;
-  };
+  ROUTES_TO_SKIP.forEach((route) => console.log(chalk.yellow(`- ${route}`)));
 
   const envConfig = getEnvironmentConfig(environment);
-  const warpRouteIds = Object.keys(registryWarpDeployConfigMap);
+  const { routesWithUnsupportedChains, warpIdsToCheck } =
+    await getWarpIdsToCheck({
+      environment,
+      envConfig,
+      registry,
+      registryWarpDeployConfigMap,
+    });
 
-  const routesWithUnsupportedChains: string[] = [];
+  logUnsupportedRoutes(routesWithUnsupportedChains);
 
-  const filterResults = await Promise.all(
-    warpRouteIds.map(async (warpRouteId) => {
-      const warpRouteConfig = registryWarpDeployConfigMap[warpRouteId];
-      const isTestnet = await isTestnetRoute(warpRouteConfig);
-      const shouldCheck =
-        (environment === 'mainnet3' && !isTestnet) ||
-        (environment === 'testnet4' && isTestnet);
-
-      if (!shouldCheck || routesToSkip.includes(warpRouteId)) {
-        return false;
-      }
-
-      const routeChains = Object.keys(warpRouteConfig);
-      const unsupportedChains = routeChains.filter(
-        (chain) => !envConfig.supportedChainNames.includes(chain),
-      );
-      if (unsupportedChains.length > 0) {
-        routesWithUnsupportedChains.push(
-          `${warpRouteId} (${unsupportedChains.join(', ')})`,
-        );
-        return false;
-      }
-
-      return true;
-    }),
+  const warpConfigChains = getWarpConfigChains(
+    warpIdsToCheck,
+    registryWarpDeployConfigMap,
   );
-
-  if (routesWithUnsupportedChains.length > 0) {
-    console.log(
-      chalk.yellow(
-        `Skipping ${routesWithUnsupportedChains.length} routes with unsupported chains:`,
-      ),
-    );
-    routesWithUnsupportedChains.forEach((route) =>
-      console.log(chalk.yellow(`  - ${route}`)),
-    );
-  }
-
-  const warpIdsToCheck = warpRouteIds.filter(
-    (_, index) => filterResults[index],
-  );
-
-  const warpConfigChains = new Set<ChainName>();
-  warpIdsToCheck.forEach((warpRouteId) => {
-    const warpRouteConfig = registryWarpDeployConfigMap[warpRouteId];
-    Object.keys(warpRouteConfig).forEach((chain) =>
-      warpConfigChains.add(chain),
-    );
-  });
 
   console.log(
     `Checking ${warpIdsToCheck.length} routes across chains: ${Array.from(warpConfigChains).join(', ')}`,
@@ -166,24 +125,11 @@ async function main() {
       if (result.violations.length > 0) {
         logWarpRouteCheckResult(result);
         if (pushMetrics) {
-          for (const violation of result.violations) {
-            checkerViolationsGauge
-              .labels({
-                actual: violation.actual,
-                chain: violation.chain,
-                contract_name: violation.name,
-                expected: violation.expected,
-                module: 'warp',
-                remote: '',
-                sub_type: '',
-                type: violation.type,
-                warp_route_id: warpRouteId,
-              })
-              .set(1);
-            console.log(
-              `Violation: ${violation.name} on ${violation.chain} with ${violation.actual} ${violation.type} ${violation.expected} pushed to metrics`,
-            );
-          }
+          pushWarpViolationsMetrics(
+            checkerViolationsGauge,
+            result,
+            warpRouteId,
+          );
         }
       } else {
         console.info(chalk.green(`warp checker found no violations`));
@@ -224,3 +170,261 @@ main()
     console.error(e);
     process.exit(1);
   });
+
+async function runWarpRouteCheckFromRegistry({
+  multiProvider,
+  warpRouteId,
+  registryUris,
+  chains,
+  warpCoreConfig,
+  warpDeployConfig,
+}: {
+  chains?: string[];
+  multiProvider: Awaited<ReturnType<EnvironmentConfig['getMultiProvider']>>;
+  registryUris: string[];
+  warpCoreConfig?: WarpCoreConfig;
+  warpDeployConfig?: WarpRouteDeployConfigMailboxRequired;
+  warpRouteId: string;
+}): Promise<WarpRouteCheckResult> {
+  const loadedConfigs = await loadWarpConfigsFromRegistry({
+    registryUris,
+    warpRouteId,
+    warpCoreConfig,
+    warpDeployConfig,
+  });
+
+  const filteredConfigs = filterWarpConfigsByChains({
+    chains,
+    warpCoreConfig: loadedConfigs.warpCoreConfig,
+    warpDeployConfig: loadedConfigs.warpDeployConfig,
+  });
+
+  return checkWarpRouteDeployConfig({
+    multiProvider,
+    warpCoreConfig: filteredConfigs.warpCoreConfig,
+    warpDeployConfig: filteredConfigs.warpDeployConfig,
+  });
+}
+
+function logWarpRouteCheckResult(result: WarpRouteCheckResult) {
+  if (Object.keys(result.diff).length > 0) {
+    console.log(chalk.yellow(yamlStringify(result.diff, null, 2)));
+  }
+
+  if (result.scaleViolations.length > 0) {
+    console.log(
+      chalk.red('Found invalid or missing scale for inconsistent decimals'),
+    );
+  }
+
+  if (result.violations.length > 0) {
+    console.table(result.violations, [
+      'chain',
+      'name',
+      'type',
+      'actual',
+      'expected',
+    ]);
+  }
+}
+
+async function loadWarpConfigsFromRegistry({
+  registryUris,
+  warpRouteId,
+  warpCoreConfig,
+  warpDeployConfig,
+}: {
+  registryUris: string[];
+  warpRouteId: string;
+  warpCoreConfig?: WarpCoreConfig;
+  warpDeployConfig?: WarpRouteDeployConfigMailboxRequired;
+}): Promise<{
+  warpCoreConfig: WarpCoreConfig;
+  warpDeployConfig: WarpRouteDeployConfigMailboxRequired;
+}> {
+  const resolvedWarpCoreConfig =
+    warpCoreConfig ??
+    (await getRegistry({
+      registryUris,
+      enableProxy: true,
+    }).getWarpRoute(warpRouteId));
+  const resolvedWarpDeployConfig =
+    warpDeployConfig ??
+    (await getWarpDeployConfigFromMergedRegistry(warpRouteId, registryUris));
+
+  assert(
+    resolvedWarpCoreConfig,
+    `Warp route config not found for ${warpRouteId}`,
+  );
+  assert(
+    resolvedWarpDeployConfig,
+    `Warp route deploy config not found for ${warpRouteId}`,
+  );
+
+  return {
+    warpCoreConfig: resolvedWarpCoreConfig,
+    warpDeployConfig: resolvedWarpDeployConfig,
+  };
+}
+
+function filterWarpConfigsByChains({
+  chains,
+  warpCoreConfig,
+  warpDeployConfig,
+}: {
+  chains?: string[];
+  warpCoreConfig: WarpCoreConfig;
+  warpDeployConfig: WarpRouteDeployConfigMailboxRequired;
+}) {
+  if (!chains?.length) {
+    return { warpCoreConfig, warpDeployConfig };
+  }
+
+  const requestedChains = new Set(chains);
+
+  const filteredWarpCoreConfig = {
+    ...warpCoreConfig,
+    tokens: warpCoreConfig.tokens.filter((token) =>
+      requestedChains.has(token.chainName),
+    ),
+  };
+  const filteredWarpDeployConfig = objFilter(
+    warpDeployConfig,
+    (chain, _config): _config is WarpRouteDeployConfigMailboxRequired[string] =>
+      requestedChains.has(chain),
+  );
+
+  assert(
+    filteredWarpCoreConfig.tokens.length > 0,
+    `None of the requested chains are present in the warp core config: ${chains.join(', ')}`,
+  );
+  assert(
+    Object.keys(filteredWarpDeployConfig).length > 0,
+    `None of the requested chains are present in the warp deploy config: ${chains.join(', ')}`,
+  );
+
+  return {
+    warpCoreConfig: filteredWarpCoreConfig,
+    warpDeployConfig: filteredWarpDeployConfig,
+  };
+}
+
+async function getWarpIdsToCheck({
+  environment,
+  envConfig,
+  registry,
+  registryWarpDeployConfigMap,
+}: {
+  environment: string;
+  envConfig: ReturnType<typeof getEnvironmentConfig>;
+  registry: ReturnType<typeof getRegistry>;
+  registryWarpDeployConfigMap: Record<
+    string,
+    WarpRouteDeployConfigMailboxRequired
+  >;
+}) {
+  const warpRouteIds = Object.keys(registryWarpDeployConfigMap);
+  const routesWithUnsupportedChains: string[] = [];
+
+  const filterResults = await Promise.all(
+    warpRouteIds.map(async (warpRouteId) => {
+      const warpRouteConfig = registryWarpDeployConfigMap[warpRouteId];
+      const isTestnet = await isTestnetRoute(registry, warpRouteConfig);
+      const shouldCheck =
+        (environment === 'mainnet3' && !isTestnet) ||
+        (environment === 'testnet4' && isTestnet);
+
+      if (!shouldCheck || ROUTES_TO_SKIP.includes(warpRouteId)) {
+        return false;
+      }
+
+      const routeChains = Object.keys(warpRouteConfig);
+      const unsupportedChains = routeChains.filter(
+        (chain) => !envConfig.supportedChainNames.includes(chain),
+      );
+      if (unsupportedChains.length > 0) {
+        routesWithUnsupportedChains.push(
+          `${warpRouteId} (${unsupportedChains.join(', ')})`,
+        );
+        return false;
+      }
+
+      return true;
+    }),
+  );
+
+  return {
+    routesWithUnsupportedChains,
+    warpIdsToCheck: warpRouteIds.filter((_, index) => filterResults[index]),
+  };
+}
+
+function logUnsupportedRoutes(routesWithUnsupportedChains: string[]) {
+  if (routesWithUnsupportedChains.length === 0) {
+    return;
+  }
+
+  console.log(
+    chalk.yellow(
+      `Skipping ${routesWithUnsupportedChains.length} routes with unsupported chains:`,
+    ),
+  );
+  routesWithUnsupportedChains.forEach((route) =>
+    console.log(chalk.yellow(`  - ${route}`)),
+  );
+}
+
+function getWarpConfigChains(
+  warpIdsToCheck: string[],
+  registryWarpDeployConfigMap: Record<
+    string,
+    WarpRouteDeployConfigMailboxRequired
+  >,
+) {
+  const warpConfigChains = new Set<ChainName>();
+  warpIdsToCheck.forEach((warpRouteId) => {
+    const warpRouteConfig = registryWarpDeployConfigMap[warpRouteId];
+    Object.keys(warpRouteConfig).forEach((chain) =>
+      warpConfigChains.add(chain),
+    );
+  });
+  return warpConfigChains;
+}
+
+async function isTestnetRoute(
+  registry: ReturnType<typeof getRegistry>,
+  warpRouteConfig: WarpRouteDeployConfigMailboxRequired,
+) {
+  for (const chain of Object.keys(warpRouteConfig)) {
+    const chainMetadata = await registry.getChainMetadata(chain);
+    if (chainMetadata?.isTestnet) {
+      return true;
+    }
+  }
+  return false;
+}
+
+function pushWarpViolationsMetrics(
+  checkerViolationsGauge: Gauge<string>,
+  result: WarpRouteCheckResult,
+  warpRouteId: string,
+) {
+  for (const violation of result.violations) {
+    checkerViolationsGauge
+      .labels({
+        actual: violation.actual,
+        chain: violation.chain,
+        contract_name: violation.name,
+        expected: violation.expected,
+        module: 'warp',
+        remote: '',
+        sub_type: '',
+        type: violation.type,
+        warp_route_id: warpRouteId,
+      })
+      .set(1);
+    console.log(
+      `Violation: ${violation.name} on ${violation.chain} with ${violation.actual} ${violation.type} ${violation.expected} pushed to metrics`,
+    );
+  }
+}
